WaveGAN_Server/requirements.txt
absl-py==0.10.0
astor==0.8.1
audioread==2.1.8
backcall==0.2.0
cffi==1.14.3
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
gast==0.2.2
gevent==20.9.0
google-pasta==0.2.0
greenlet==0.4.17
grpcio==1.32.0
h5py==2.10.0
importlib-metadata==2.0.0
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
joblib==0.17.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
librosa==0.6.2
llvmlite==0.31.0
Markdown==3.2.2
matplotlib==3.0.2
numba==0.48.0
numpy==1.18.5
opt-einsum==3.3.0
parso==0.7.1
pickleshare==0.7.5
prompt-toolkit==3.0.7
protobuf==3.13.0
pycparser==2.20
Pygments==2.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
resampy==0.2.2
scikit-learn==0.23.2
scipy==1.1.0
SimpleWebSocketServer==0.1.1
six==1.15.0
sounddevice==0.4.1
tensorboard==1.15.0
tensorflow-cpu-estimator==1.15.1
tensorflow-cpu==1.15.0
termcolor==1.1.0
threadpoolctl==2.1.0
traitlets==4.3.3
wcwidth==0.2.5
websocket-client-py3==0.15.0
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.3.0
zope.event==4.5.0
zope.interface==5.1.2
pydub==0.25.1

SoundFile~=0.10.3.post1